Extract duplicate tests helpers into a shared scope
