+++ EssentialDeveloper/EssentialDeveloperTests/Feed Cache/CacheFeedUseCaseTests.swift
@@ -121,23 +121,4 @@ final class CacheFeedUseCaseTests: XCTestCase {
         wait(for: [exp], timeout: 1.0)
         XCTAssertEqual(receivedError as NSError?, expectedError, file: file, line: line)
     }
-    
-    private func uniqueImage() -> FeedImage {
-        return FeedImage(id: UUID(), description: "any", location: "any", url: anyURL())
-    }
-    
-    private func uniqueImageFeed() -> (models: [FeedImage], local: [LocalFeedImage]) {
-        let models = [uniqueImage(), uniqueImage()]
-        let localItems = models.map {
-            LocalFeedImage(id: $0.id, description:  $0.description, location: $0.location, url: $0.url)}
-        return (models, localItems)
-    }
-
-    private func anyURL() -> URL {
-        URL(string: "https://any-url.com")!
-    }
-    
-    private func anyNSError() -> NSError {
-        return NSError(domain: "any eror", code: 0)
-    }
 }
