+++ EssentialDeveloper/EssentialDeveloperTests/Feed Cache/Helpers/FeedCacheTestHelpers.swift
@@ -1,0 +1,23 @@
+import Foundation
+import EssentialDeveloper
+
+func uniqueImage() -> FeedImage {
+    return FeedImage(id: UUID(), description: "any", location: "any", url: anyURL())
+}
+
+func uniqueImageFeed() -> (models: [FeedImage], local: [LocalFeedImage]) {
+    let models = [uniqueImage(), uniqueImage()]
+    let localItems = models.map {
+        LocalFeedImage(id: $0.id, description:  $0.description, location: $0.location, url: $0.url)}
+    return (models, localItems)
+}
+
+extension Date {
+    func adding (days: Int) -> Date {
+        return Calendar(identifier: .gregorian).date(byAdding: .day, value: days, to: self)!
+    }
+    
+    func adding (seconds: TimeInterval) -> Date {
+        return self + seconds
+    }
+}
