+++ EssentialDeveloper/EssentialDeveloperTests/Feed Cache/LoadFeedFromCacheTests.swift
@@ -167,33 +167,4 @@ final class LoadFeedFromCacheTests: XCTestCase {
         wait(for: [exp], timeout: 1.0)
         
     }
-    
-    private func anyNSError() -> NSError {
-        return NSError(domain: "any eror", code: 0)
-    }
-    
-    private func anyURL() -> URL {
-        URL(string: "https://any-url.com")!
-    }
-    
-    private func uniqueImage() -> FeedImage {
-        return FeedImage(id: UUID(), description: "any", location: "any", url: anyURL())
-    }
-    
-    private func uniqueImageFeed() -> (models: [FeedImage], local: [LocalFeedImage]) {
-        let models = [uniqueImage(), uniqueImage()]
-        let localItems = models.map {
-            LocalFeedImage(id: $0.id, description:  $0.description, location: $0.location, url: $0.url)}
-        return (models, localItems)
-    }
-}
-
-private extension Date {
-    func adding (days: Int) -> Date {
-        return Calendar(identifier: .gregorian).date(byAdding: .day, value: days, to: self)!
-    }
-    
-    func adding (seconds: TimeInterval) -> Date {
-        return self + seconds
-    }
 }
