+++ EssentialDeveloper/EssentialDeveloperTests/Helpers/SharedTestsHelpers.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+func anyURL() -> URL {
+    URL(string: "https://any-url.com")!
+}
+
+func anyNSError() -> NSError {
+    return NSError(domain: "any eror", code: 0)
+}
